+++ CHANGES.md
@@ -1,3 +1,12 @@
+### 1.1.0 ###
+* Added: restart delay which is dynamically increased by the number of failed processes
+* Added: SIGINT handler
+* Added: simple example
+* Added: promise based contol flow
+* Added: minimalistic example
+* Changed: some logging levels to notice/alert
+* Changed: separated cluster startup, worker management and hot-reload logic
+
 ### 1.0.2 ###
 * Bugfix: Invalid module export caused by previous renaming